抽象语法树节点文件结构调整
