+++ metasequoia_sql/core/node/__init__.py
@@ -2,5 +2,6 @@
 抽象语法树（AST）节点
 """
 
+from metasequoia_sql.core.node.sql_type import *
 from metasequoia_sql.core.node.objects import *
